Update NESTML to v8.0.0 (#254)
* feat: add devcontainer and Dependabot configuration

Add a devcontainer configuration for an Ubuntu environment 
with Docker support. Include features for Docker-in-Docker, 
Docker Buildx, and Docker Compose. Introduce a Dependabot 
configuration to automate dependency updates for 
devcontainers on a weekly basis. These changes enhance 
development efficiency and ensure up-to-date dependencies.

* fix: update nestml to version 8.0.0 in Dockerfiles

Update the nestml package version from 8.0.0-rc3 to 8.0.0 in bothDockerfiles. change ensures that the latest stable version of nestml is used, improving compatibility and stability of the 
environment.