Integrate upstream changes and resolve Docker configuration conflicts (#263)
* Merge build and deploy to fix deploy problems

* Add nest-server requirements

* Fix nest-desktop and compose file

* Add v3.7

* Remove test

* Add v3.7 to readme

* Fix nestml version and add pygsl

* Install nestml fix for nest verssion

* Fix merging errors

* Deactivate build of 3.7 and 2.20.2 in CI

* Add rc2

* NEW Version 3.8

* Fix the mermaid installation with a patch and fix some doubles

* Fix numpy installation

* Add NESTML v8.0.0rc3

* feat: add devcontainer and Dependabot configuration

Add a devcontainer configuration for an Ubuntu environment 
with Docker support. Include features for Docker-in-Docker, 
Docker Buildx, and Docker Compose. Introduce a Dependabot 
configuration to automate dependency updates for 
devcontainers on a weekly basis. These changes enhance 
development efficiency and ensure up-to-date dependencies.

* fix: update nestml to version 8.0.0 in Dockerfiles

Update the nestml package version from 8.0.0-rc3 to 8.0.0 in bothDockerfiles. change ensures that the latest stable version 
of nestml is used, improving compatibility and stability of the 
environment.

* Delete devcontainer file

* Remove  double of pip install numpy in Dockerfile

* Comment out Docker registry login in CI configuration

* fix: update destination repository URLs in ebrains-push workflow